removed reference to v2 from agency unit tests
