+++ tests/containers/agency/visualizations/ObligationsByAwardTypeContainer-test.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { render, waitFor } from 'test-utils';
 import ObligationsByAwardTypeContainer from 'containers/agency/visualizations/ObligationsByAwardTypeContainer';
 import * as agencyV2 from 'apis/agency';
-import * as actions from 'redux/actions/agency/agencyV2Actions';
+import * as actions from 'redux/actions/agency/agencyActions';
 import { defaultState } from '../../../testResources/defaultReduxFilters';
